feat: backend connections tab
