+++ frontend/src/component/admin/network/NetworkTrafficUsage/bar-chart-options.ts
@@ -1,0 +1,102 @@
+import type { Theme } from '@mui/material/styles/createTheme';
+import type { ChartOptions } from 'chart.js';
+import { formatTickValue } from 'component/common/Chart/formatTickValue';
+
+export const createBarChartOptions = (
+    theme: Theme,
+    tooltipTitleCallback: (tooltipItems: any) => string,
+    includedTraffic?: number,
+): ChartOptions<'bar'> => ({
+    plugins: {
+        annotation: {
+            clip: false,
+            annotations: {
+                line: {
+                    type: 'line',
+                    borderDash: [5, 5],
+                    yMin: includedTraffic ? includedTraffic / 30 : 0,
+                    yMax: includedTraffic ? includedTraffic / 30 : 0,
+                    borderColor: 'gray',
+                    borderWidth: 1,
+                    display: !!includedTraffic,
+
+                    label: {
+                        backgroundColor: 'rgba(192, 192, 192,  0.8)',
+                        color: 'black',
+                        padding: {
+                            top: 10,
+                            bottom: 10,
+                            left: 10,
+                            right: 10,
+                        },
+                        content: 'Average daily requests included in your plan',
+                        display: !!includedTraffic,
+                    },
+                },
+            },
+        },
+        legend: {
+            position: 'bottom',
+            labels: {
+                color: theme.palette.text.primary,
+                pointStyle: 'circle',
+                usePointStyle: true,
+                boxHeight: 6,
+                padding: 15,
+                boxPadding: 5,
+            },
+        },
+        tooltip: {
+            backgroundColor: theme.palette.background.paper,
+            titleColor: theme.palette.text.primary,
+            bodyColor: theme.palette.text.primary,
+            bodySpacing: 6,
+            padding: {
+                top: 20,
+                bottom: 20,
+                left: 30,
+                right: 30,
+            },
+            borderColor: 'rgba(0, 0, 0, 0.05)',
+            borderWidth: 3,
+            usePointStyle: true,
+            caretSize: 0,
+            boxPadding: 10,
+            callbacks: {
+                title: tooltipTitleCallback,
+            },
+        },
+    },
+    responsive: true,
+    scales: {
+        x: {
+            stacked: true,
+            ticks: {
+                color: theme.palette.text.secondary,
+            },
+            grid: {
+                display: false,
+            },
+        },
+        y: {
+            stacked: true,
+            ticks: {
+                color: theme.palette.text.secondary,
+                maxTicksLimit: 5,
+                callback: formatTickValue,
+            },
+            grid: {
+                drawBorder: false,
+            },
+        },
+    },
+    elements: {
+        bar: {
+            borderRadius: 5,
+        },
+    },
+    interaction: {
+        mode: 'index',
+        intersect: false,
+    },
+});
